Create logspout-papertrail image to communicate over TLS
In order to create a logspout container to communicate over TLS to
papertrail this commit creates a `Dockerfile` to build a
`logspout-papertrail` docker image.  This commit creates a `Dockerfile`
based on the [Custom Logspout
Builds](https://github.com/gliderlabs/logspout/tree/master/custom) with
the papertrail cert overwriting the CA-Certs on the image, making this
image exclusively only able to communicate over TLS to papertrail.  This
commit also updates the `README.md` for directions on using the image.

[74726600553133](https://app.asana.com/0/74726600553133/74726600553133)